completing for user story 1.5
